refactor(agents): build the prompt date line from one helper
todaysDatePromptLine() now owns the whole sentence, and the date formatter
is private to that module, so no call site can assemble its own "Today's
date:" prefix and drift again.

The line states its own format — "Today's date: 2026-09-02 (Date format
YYYY-MM-DD)" — because a bare ISO date still reads as day-first to some
models.
