+++ apps/api/src/external/llm/providers/live-regressions/fat-prompt.fixture.ts
@@ -395,5 +395,5 @@ ${toolsSection}
 ## Response language:
 Always answer in English.
 
-Today's date: 2026-07-29`
+Today's date: 2026-07-29 (Date format YYYY-MM-DD)`
 }
